scrapers/requirements-api.txt
# API-Only Scraper Requirements
# Much simpler than the original! No database dependencies needed.

# HTTP requests (for API calls)
requests>=2.31.0

# Web scraping (for your existing scrapers)
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Environment variables
python-dotenv>=1.0.0

# Optional: Advanced scraping (only if you use these in your scrapers)
selenium>=4.15.0
playwright>=1.40.0

# Note: No more supabase, no more database dependencies!
# The API handles everything now. 